requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
build==1.0.3
CacheControl==0.13.1
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
crashtest==0.4.1
distlib==0.3.8
Django==5.0.3
django-debug-toolbar==4.3.0
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dulwich==0.21.7
fastjsonschema==2.19.1
filelock==3.13.1
idna==3.6
importlib-metadata==7.0.1
installer==0.7.0
jaraco.classes==3.3.1
keyring==24.3.0
kombu==5.3.5
more-itertools==10.2.0
msgpack==1.0.7
packaging==23.2
pexpect==4.9.0
pkginfo==1.9.6
platformdirs==3.11.0
poetry==1.7.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pycparser==2.21
PyJWT==2.8.0
pyproject_hooks==1.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
rapidfuzz==3.6.1
redis==5.0.1
requests==2.31.0
requests-toolbelt==1.0.0
shellingham==1.5.4
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.12.3
trove-classifiers==2024.1.31
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
vine==5.1.0
virtualenv==20.25.0
wcwidth==0.2.13
xattr==0.10.1
zipp==3.17.0
